fix: react 18 Strict mode issues
The one behavioral change made is that The Router component will no longer `dispose` of the farce navigation subscription on unmount. This never was correct to do in the first place since remounting would not re subscribe. If you need to tear down the farce protocol you can dispatch dispose manually